Add API utility functions and refactor beamline.py for improved structure
